Added new function changeBackgroundColor()
The changeBackgroundColor() function contains a conditional statement that changes the background color to 4 different colors based on where ballX and ballY coordinates are in the window. If the ball is in the upper left quadrant, the color will be peach orange. If it is in the upper right quadrant, the color will be bubblegum pink. If it is in the bottom left quadrant,  the colour will be sky blue, and if the ball is in the bottom right quadrant, the background color will be pale green.