exercise03: added mouseClicked() in main program
If a bouncer is clicked, it will get bigger.  If you click on the other bouncer that one will get bigger, and it was also result in returning the other bouncer back to its default size. You can also click anywhere in the window and it will return the bouncers to their default size.