midterm: CHANGED isOffScreen()
Added an if() statement to check if the ball is off the left or right side of the window. If the ball goes off the left side, the right player's score, rightScore, will increase by 1, rightScore++...
If the ball goes off the right side, the left player's score, leftScore, will increase by 1, leftScore++. If the ball goes off, the function will return true, and if it does not, it will return false.