midterm: CHANGED background color
The background color now displays a continuous cycle of different colors with a smooth transition from color to color. I changed the color mode to HSB, and I added a variable (r) to hold the floating-point number for the R value in RGB. As long as the r is not equal to or greater than 255, it will keep increasing by 1, and its value will be called in the background(r,255,255). Once r is >= 255, r will reset to 0, and continue again by adding 1.