midterm: FIXED scrolling text and game over
The scrolling text will scroll through the window only if the game over IS NOT prompted.  Once the game over is prompted, the scrolling text will disappear, and once the game is started again, the scrolling text's index will start at 0, and it will be scrolling through the window again.
For the whoWins function, I had to add in keyPressed along with the keyCode in order for it to only work if it is being pressed, and I had to change ENTER to 0 because there are issues with how keyCode behaves across different renderers and operating systems. So, I have to use the number value for the enter key, which is 0.