exercise03: CHANGED mouseClicked(), handleMouse() and variable names, and edited comments
In mouseClicked() and in handleMouse(), I changed  (dist(mouseX, mouseY, x, y) < size/2) to (dist(mouseX, mouseY, x, y) < size).  I wanted to check for the whole size rather than just half the size so there is more of a range to get it on the ball.  I changed mousePressedFillColor to mouseClickedFillColor, and I changed tempMousePressedFillColor to tempMouseClickedFillColor because "clicked" is the more appropriate term to use for what is happening, rather than "pressed".