midterm: ADDED collidesWithBlocker() method.
This method checks whether the ball is colliding with the blocker passed as an argument. If it is, it makes the ball bounce away from the blocker by reversing its x velocity. The method contains variables that calculate possible overlaps with the blocker on all sides. The if statement checks if the ball overlaps with the blocker.