midterm: ADDED changeControlKeys() function. and ADDED in the resets().
This function changes the control keys for the players.  If the left score is more than or equal to half the winning score value, the right player's up and down control keys will switch to other keys. If the right score is more than or equal to half the winning score value, the left player's up and down control keys will switch to other keys. In the resets() function, the control keys for both players will return to their default values.