midterm: ADDED how the players are prompted to change control keys
When one player reaches half of the winning score, the other player will be prompted to change their control keys. rameRateCountLeft and rameRateCountRight variables were added to act as counters for the amount of time the text is displayed for. The text will be displayed for 100 frames. Once the counters reach 100 frames, they will reset in the if (!gameOver) conditional.