exercise04: ADDED conditionals for the griddies
These conditionals are for changing the griddie's color and size.  If a griddie collides with a cyan dottie, then that griddie will become orange and it will get bigger.  If a griddie collides with a magenta dottie, then that griddie will become lime green and it will get bigger.  If a griddie collides with a yellow dottie, then that griddie will become white and it will get bigger.