final_project: Created class called Marine
This Marine class is basically the planet protector/superhero character, and the hero's name is Marine.  In this class there are Marine's properties: its position and velocity, its speed, and its width and height.  In the constructor is where we set Marine's position and controls based on arguments, and its initial x velocity is 0.