final_project: Added keyPressed() and keyReleased()
The keyPressed() checks if the LEFT or RIGHT keys are being pressed.  If the left key is pressed, the x velocity is given a negative speed to move towards the left.  If the right key is being pressed, the x velocity is given the positive speed to move towards the right.
The keyReleased() checks if the left and right keys are released, and if they are, the x velocity will be set to 0 so Marine can stop moving.
These methods are called in the main program.