exercise07: ADDED update() and display() methods + ADDED comments
update() --- The bubble moves downwards on the y-axis.  If the y coodinate is greater than the height of the window + size of the bubble,  then the y coordinate is given the negative value of the bubble size and basically goes above the window, and comes downwards again with a random velocity and a random x coordinate as well.

display() --- the bubble is created with no stroke, it is given the random array fill color and it is created by using an ellipse