exercise07: ADDED playSound() method
playSound() --- basically to check conditions to figure out which sound to play.  If the mouse is over a rose colored bubble, sound01 will play.  If the mouse is over a mustard yellow colored bubble, sound02 will play.  If the mouse is over a mint green colored bubble, sound03 will play.