exercise07: Created a Griddie class
Same code as Bubble class with a few changes...
The griddie moves along the x-axis, so the vy is changed to vx.

update() --- the conditional checks if the griddie has left the window by checking if the x value if greater than the width + the size of the griddie, and if it is, the x gets updated with the vx, the y value is given a random value within the height of the window and the vx is assigned a random number between 3-7.

display() --- uses rect, and uses the random fill colors or white and violet.

playSound() --- if a white griddie appears, the kick sound will play. If a violet griddie appears, the snare sound will play.