exercise07: Added Griddie[] array + more code
Griddie[] --- an array of griddies, like the bubbles array but they are squares instead and move along the x-axis.
Also created a griddiesCounter variable to keep track of the griddies being displayed.

SoundFile --- added two new sound files, kickSound and snareSound.

setup() --- the for-loop creates the amount of griddies that are stored in the array.

draw() --- the for-loop goes through the griddiesCounter and updates the griddie movement and displays that griddie being referenced.

keyPressed() --- here is where the griddiesCounter gets incremented. Also, depending on which griddie is being displayed (white or violet), the kick or snare sound will play.  Once the griddiesCounter EQUALS the griddies array length, the griddies counter will become the value of the length -1 to it can always keep adding more griddies to be displayed.