exercise08: Added the update() and display() methods
update() --- The star moves downwards on the y-axis.  If the y coodinate is greater than the height of the window + size of the star,  then the y coordinate is given the negative value of the star size and basically goes above the window, and comes downwards again with a random velocity and a random x coordinate as well.

display() --- the star is created with no stroke, it is given the random array fill color and it is created by using an ellipse