exercise08: Added the Starfield array in the main program.
Added a for-loop in the setup() to loop through and create all the stars in the array one by one.  Added a displayStarfieldBackground() function:  added a for-loop to loop through all the stars one by one and update the stars and display them.  The displayStarfieldBackground() function is called in the draw() function.