exercise08: edited update() and display() methods
Removed assigning random value to vy, as it was reduntant since I assign the random value to vy in the main program.  Created a size2 variable and assign it the same value as size, so the width and height of the ellipse can be the same values, instead of assigning different random values to both like what was happening before.