final_project: Created the update() and display() methods
update() --- Where the star gets moved.  We add its current y-velocity value to its y-position. Also checks if the star has exited the window by the bottom, and if so, the y position is given the negative size value, so it is placed right above the top of the window and comes back down. The x value is assigned a random number withint the width of the window.

display() --- Where the star gets created. Since the size gets assigned random size values from the main program, I had to create a starRadius variable. I assign the value of size to the starRadius to ensure that the width and height of the circle are the same value to make a perfect circle, rather than making an oval with differentm random numbers for width and height.