final_project: Added StarBackground[] & created displayStarBackground()
Created the array storing all the stars for the starBackground.

setup() --- Created the for-loop to create the stars that are stored in the starBackground array.

displayStarBackground() --- Created a for-loop to loop through all the stars one by one, and to update  and display each star.

display() --- Call the displayStarBackground() function to update and display the stars.