final_project: Added Villain class
Villain --- sets the position, speed and size based on arguments from the main program. Assign the vx and vy the valued of the speed.

update() --- the villain moves from left to right, so the current vx value is added to the x value.

hitsWall() --- when the villain hits the left or right side of the window, its vx becomes negative, so it can go the opposite direction.  the y value gets assigned the vallue of half the size of the villain, and this makes the villain shift downwards.

dies() --- when a villain gets hit and dies, it's energy (alpha channel) will become 0 and it will dissapear.

display() --- creates the villain in the shape of an ellipse, with no stroke and a fill that contains a random color array.