final_project: Added villains array
in setup() --- Added a for-loop to create the amount of villains that are stored in the villains array.  Assigned values for the villain properties.

in draw() --- Created a boolean variable hitsEdge, that is set to false to start off with. Created a for-loop to loop through the villains one by one to update and display them.  In this loop, we also check if the indexed villain hits the right edge OR left edge of the window.  If it does, hitsEdge becomes true.  Outside of this loop we have an if statement checking if hitsEdge is true, and it if is true, then we loop through each villain again and the referenced villain calls the hitsWall() function form the villain class.