final_project: Added comments + code for removing a laser once it hits a villain
in the for-loop where the lasers are being created, the disappear() method gets called in the if statement that is checking if a villain has been hit.

the getRemoved boolean variable is being checked in the if-statement that is in the new for-loop that is removing lasers once theyve hit their target. This for-loop is being looped through backwards, and removing the laser being referenced.