final_project: Added ArrayList + created shootsLaser() method
Created a new array list, villainLaser.  We create the villainLaser at a frandom framerate in the new shootsLaser() method.  The villain lasers are updated and displayed one by one in a for-loop that loops through the size of the array list.  the lasers come out from the x and y positions of the villains.  A laser gets removed from the array list once it collides with the planet, the superhero or the shield. this is checked in the if statements that checks the distances between the lasers and the planet, superhero and shield.