final_project: Added array for 2nd row of villains. Added code for shield...
Created another array for the 2nd row of villains. Also created a hitsEdge2 boolean variable to fix problem with adding too much space between the villain rows.

Created shieldOn boolean to check if the shield is displayed or not. It if set to true once the 's' key is pressed, and it is set back to false once the 's' key is released.  if the shield is displayed, the superhero's speed is set to 0, so the superhero can only move when it is not using its shield.  once the shield is away, the superhero's speed gets set back to its default speed.