final_project: Added functions for main menu
I control which screen is active by updating the gameScreen variable. I display the correct screen according to the value of this variable. I define a different method for each screen to display. In the draw(), it simply checks the value of the gameScreen variable, and calls the corresponding method. in mousePressed(), it is listening to mouse clicks and if the active screen is 0, the initial screen, it will call the startGame(), which starts the game