final_project: Added start screen + game over screen and created PImage ArrayList<>
the correct screen is called according to the value of the gameScreen variable.  In the draw() there is an if statement checking the gameScreen value: if it is 0 the startScreen function will be called and the contents will be displayed, if it is 1 the gameScreen function will be called and the contents will be displayed, and if it is 2 the gameOverScreen function will be called and the contents will be displayed.  the gameScreen starts off at 0 from the starts so the startScreen content can show up first.

PImage ArrayList --- created an ArrayList called hearts to be filled with image objects: heart1, heart2, heart3.  in the setup() heart1, heart2 and heart3 are loaded with the heart PNGs, and then these variables are added in the hearts ArrayList.
displayHearts() --- the heart images in the hearts ArrayList are created in this function and being placed in the header, one heart next to the other.