final_project: Added planet object, edited hearts, and created resetGame() function
Planet planet --- created global variable for the planet.  In the draw() function, the planet is updated and displayed.

heartsSize --- is set to 3 for the 3 initial hearts being displayed. this variable is then used to loop through and display the heart images.

resetGame() --- this function is called when the user clicks in the gameOverScreen.  it essentially resets the superhero to its initial position, resets the villains to bring them back to the top, the planet is brought back to its full life, and the hearts are restored.