final_project: Villain laser + planet collision... Villain laser + shield collision... added reachesSuperhero() method
Villain laser + planet collision --- Whenever a villain laser hits the planet, not only will the villain laser get removed, but the alpha channel for the planet will decrement by 15 each time.  the planet will lose life every time it gets hit, and when this eventually results in the planet disappearing, the user loses.

reachesSuperhero() --- whenever the villains reach the superhero, the user loses and gameScreen is given the value 2, which will display the game over screen.