final_project: Created a Meteor class for the pong ball for level 2
This meteor will act as the pong ball for level 2, where the user not only has to protect the planet from the villains, but also keep the meteor from hitting the planet.

collide() --- the first if-statement checks for collision with the shield. If it hits the shield, the y velocity will be negative and it will bounce off the shield going upwards. the second if-statements checks for collisions with the left, right and top walls of the screen, and checks if the ball exits by the bottom.  the ball will bounce off the top, right and left walls. if it exits the bottom without touching the planet, it will just reset and spawn from the top again.  the third if-statement checks if the meteor has hit the planet, and if it has, the user loses.