final_project: Added comments... Added more code.... Re-arranged code...
Added global variables for the StartScreen, GameOverScreen and PlayerWinsScreen, Level1Resets, and Level2Resets classes, so their display() methods can be called from the main program.

Meteor --- added the global variable for the meteor for level 2... created a boolean variable, meteorOn, and it is initially set to FALSE because it will only be set to TRUE when the user plays level 2... if the meteor is TRUE, its update() display() and collide() methods are called.